+++ source/ensure/EnsureDiagnosticText.ts
@@ -1,0 +1,5 @@
+export class EnsureDiagnosticText {
+  static argumentMustBeProvided(argumentNameText: string): string {
+    return `An argument for '${argumentNameText}' must be provided.`;
+  }
+}
