+++ source/tstyche.ts
@@ -2,6 +2,7 @@ export * from "#cli";
 export * from "#collect";
 export * from "#config";
 export * from "#diagnostic";
+export * from "#ensure";
 export * from "#environment";
 export * from "#events";
 export * from "#expect";
